+++ AlchemistCode/Relics/WeatheredKit.cs
@@ -1,12 +1,8 @@
-using MegaCrit.Sts2.Core.Commands;
-using MegaCrit.Sts2.Core.Entities.Creatures;
 using MegaCrit.Sts2.Core.Entities.Players;
 using MegaCrit.Sts2.Core.Entities.Relics;
 using MegaCrit.Sts2.Core.Entities.RestSite;
 using MegaCrit.Sts2.Core.HoverTips;
 using MegaCrit.Sts2.Core.Models;
-using MegaCrit.Sts2.Core.Models.Potions;
-using MegaCrit.Sts2.Core.Rooms;
 
 namespace Alchemist.AlchemistCode.Relics;
 
@@ -19,13 +15,6 @@ public class WeatheredKit : AlchemistRelic
     // Without this, BaseLib falls back to Circlet for the Touch of Orobas starter upgrade
     public override RelicModel? GetUpgradeReplacement() => ModelDb.Relic<GildedKit>();
 
-    public override async Task AfterPotionUsed(PotionModel potion, Creature? target)
-    {
-        if (potion.Owner != Owner) return;
-        if (potion is FoulPotion && Owner.RunState.CurrentRoom is MerchantRoom) return;
-        Flash();
-        await CreatureCmd.Heal(Owner.Creature, 3m);
-    }
 
     public override bool TryModifyRestSiteOptions(Player player, ICollection<RestSiteOption> options)
     {